+++ vscode-extension/src/testEntry.ts
@@ -19,7 +19,7 @@ export {
   MAX_DOCUMENT_BYTES
 } from './workflowDocument';
 export { RevisionLineage, canonicalJson, jsonDepth, MAX_JSON_DEPTH, lenientRevision, semanticJson, BoundedMap, BoundedSet } from './revisionLineage';
-export { displayText, displayIssue } from './displayText';
+export { displayText, displayIssue, INVISIBLE_RANGES } from './displayText';
 export { normCase, identity, DependencySet } from './fileIdentity';
 export { buildRefinementPrompt, toPosixRelative, escapeJsonText, REFINE_INTENTS } from './refinePrompt';
 export {
@@ -27,6 +27,7 @@ export {
   ReloadGeneration,
   ValidationScheduler,
   readArtifactFile,
+  folderSpelling,
   AUTHORED_VIEW_TYPE,
   OPEN_AUTHORED_COMMAND
 } from './authoredPanel';